requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiohttp-retry==2.8.3
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.2.0
azure-core==1.30.1
azure-identity==1.16.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.1
bs4==0.0.2
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chainlit==1.1.306
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==42.0.5
dataclasses-json==0.5.14
decorator==5.1.1
defusedxml==0.7.1
demjson3==3.0.6
Deprecated==1.2.14
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
docopt==0.6.2
docstring-parser==0.15
docx2txt==0.8
executing==2.0.1
fastapi==0.110.2
fastapi-socketio==0.0.10
fastjsonschema==2.19.1
filelock==3.13.4
filetype==1.2.0
Flask==3.0.3
Flask-Cors==4.0.0
flatbuffers==24.3.25
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.3.1
google-api-core==2.19.1
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.56.4
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.25.2
httpx-sse==0.4.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.0
importlib_resources==6.4.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kubernetes==29.0.0
lancedb==0.3.6
Lazify==0.4.0
literalai==0.0.607
llama-index==0.10.29
llama-index-agent-openai==0.1.7
llama-index-cli==0.1.11
llama-index-core==0.10.29
llama-index-embeddings-azure-openai==0.1.6
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-azure-openai==0.1.5
llama-index-llms-openai==0.1.15
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.19
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.0
llamaindex-py-client==0.1.18
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.1.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.16.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pg8000==1.31.1
pgvector==0.2.5
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.1
portalocker==2.8.2
posthog==3.5.0
prettytable==3.10.0
prompt-toolkit==3.0.36
proto-plus==1.23.0
protobuf==3.20.0
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
py==1.11.0
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.2.1
pydantic_core==2.20.1
Pygments==2.17.2
PyJWT==2.8.0
pylance==0.8.21
pymemgpt==0.3.19
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pyparsing==3.1.2
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.0.0
python-box==7.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-multipart==0.0.9
python-socketio==5.11.2
pytz==2023.4
PyYAML==6.0.1
pyzmq==26.0.3
questionary==2.0.1
ratelimiter==1.2.0.post0
redis==5.0.3
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scramp==1.4.5
semver==3.0.2
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sqlalchemy-json==0.7.0
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.16
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.37.2
striprtf==0.0.26
sympy==1.12
syncer==2.0.3
tenacity==8.2.3
tiktoken==0.5.2
tinycss2==1.3.0
tokenizers==0.15.2
tomli==2.0.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
twilio==9.0.5
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uptrace==1.24.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.20.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.2
wrapt==1.16.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.4
zep-python==2.0.0rc5
zipp==3.18.1
